requirements.txt
python-telegram-bot==21.10
requests==2.31.0
rapidfuzz==3.6.1
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
beautifulsoup4==4.12.2
